compatibility fix with kqemu-x86_64
git-svn-id: svn://svn.savannah.nongnu.org/qemu/trunk@1392 c046a42c-6fe2-441c-8c8c-71466251a162

+++ exec.c
@@ -80,7 +80,7 @@ typedef struct PageDesc {
 
 typedef struct PhysPageDesc {
     /* offset in host memory of the page + io_index in the low 12 bits */
-    unsigned long phys_offset;
+    uint32_t phys_offset;
 } PhysPageDesc;
 
 typedef struct VirtPageDesc {
